Thu, May 23, 2024, 9:32 PM +07:00

+++ finance/app.py
@@ -389,7 +389,7 @@ def deposit():
         if not inputs.isdigit():
             flash("You must input number only.")
             return redirect("/deposit")
-        amount = int(inputs)
+        amount = float(inputs)
         # Ensure password and confirmation matched
         if amount < 100:
             return apology("Minimum amount deposit: $100", 403)
